requirements.txt
discord.py>=2.4.0,<3.0.0
asyncpg>=0.30.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
